finished initial refactor, before real rewrites
