+++ source/js/buyers-guide/search/continue.md
@@ -6,8 +6,8 @@
 - [x] continue the refactor
 - [x] move all "object functions" in search over as (temporary) "class functions"
 - [x] move all "can be moved to utils" functions into their own (possibly more than one) util modules
-- [ ] move the setupXYZ functions back into the class (where sensible)
-- [ ] we should be done with the initial refactor now
+- [x] move the setupXYZ functions back into the class (where sensible)
+- [x] we should be done with the initial refactor now
 
 
 Outstanding tasks:
