+++ source/js/buyers-guide/search/search-filter.js
@@ -1,37 +1,94 @@
-import {
-  setupNavLinks,
-  setupGoBackToAll,
-  setupPopStateHandler,
-  performInitialHistoryReplace,
-  setupSearchBar,
-} from "./member-functions.js";
-
 import { Utils } from "./utils.js";
 import { CreepUtils } from "./creep-utils.js";
 import { markScrollStart } from "./slider-area.js";
-
-const categoryTitle = document.querySelector(`.category-title`);
-const parentTitle = document.querySelector(`.parent-title`);
-const subcategories = document.querySelectorAll(`.subcategories`);
-const subContainer = document.querySelector(`.subcategory-header`);
+import { setupHistoryManagement, applyHistory } from "./history.js";
+import { setupNavLinks, setupGoBackToAll } from "./member-functions.js";
 
 /**
  * ...
  */
 export class SearchFilter {
   constructor() {
+    const { searchBar, searchInput } = this.setupSearchBar();
+    setupNavLinks(this);
+    setupGoBackToAll(this);
+    setupHistoryManagement(this, searchBar, searchInput);
+
+    const subContainer = document.querySelector(`.subcategory-header`);
     [`mousedown`, `touchstart`].forEach((type) =>
       subContainer.addEventListener(type, markScrollStart)
     );
-    this.setup();
+
+    this.allProducts = document.querySelectorAll(`figure.product-box`);
+    this.categoryTitle = document.querySelector(`.category-title`);
   }
 
-  setup() {
-    const { searchBar, searchInput } = setupSearchBar(this);
-    setupNavLinks(this, searchBar, searchInput);
-    setupGoBackToAll(this, searchBar, searchInput);
-    setupPopStateHandler(this, searchBar, searchInput);
-    performInitialHistoryReplace(this, searchBar, searchInput);
+  /**
+   * Set up the search filter functionality, and return the
+   * searchbar and searchinput elements for external function
+   * binding purposes.
+   */
+  setupSearchBar() {
+    const searchBar = (this.searchBar = document.querySelector(
+      `#product-filter-search`
+    ));
+
+    if (!searchBar) {
+      return console.warn(
+        `Could not find the PNI search bar. Search will not be available.`
+      );
+    }
+
+    const searchInput = (this.searchInput = searchBar.querySelector(`input`));
+
+    searchInput.addEventListener(`input`, (evt) => {
+      const searchText = searchInput.value.trim();
+
+      if (searchText) {
+        searchBar.classList.add(`has-content`);
+        this.filter(searchText);
+      } else {
+        this.clearText();
+        applyHistory(this);
+      }
+    });
+
+    const clear = searchBar.querySelector(`.clear-icon`);
+    if (!clear) {
+      return console.warn(
+        `Could not find the PNI search input clear icon. Search will work, but clearing will not.`
+      );
+    }
+
+    clear.addEventListener(`click`, (evt) => {
+      evt.preventDefault();
+      searchInput.focus();
+      this.clearText();
+      applyHistory(this);
+    });
+
+    return { searchBar, searchInput };
+  }
+
+  /**
+   * Clear the search text
+   */
+  clearText() {
+    const { searchBar, searchInput } = this;
+    searchBar.classList.remove(`has-content`);
+    searchInput.value = ``;
+
+    this.allProducts.forEach((product) => {
+      product.classList.remove(`d-none`);
+      product.classList.add(`d-flex`);
+    });
+
+    CreepUtils.sortOnCreepiness();
+    CreepUtils.moveCreepyFace();
+
+    const state = { ...history.state, search: "" };
+    const title = Utils.getTitle(this.categoryTitle.value.trim());
+    history.replaceState(state, title, location.href);
   }
 
   /**
@@ -50,7 +107,7 @@ export class SearchFilter {
     Utils.toggleProducts(text);
 
     const state = { ...history.state, search: text };
-    const title = Utils.getTitle(categoryTitle.value.trim());
+    const title = Utils.getTitle(this.categoryTitle.value.trim());
     history.replaceState(state, title, location.href);
 
     Utils.sortFilteredProducts();
@@ -59,14 +116,15 @@ export class SearchFilter {
   }
 
   clearCategories() {
-    this.filterCategory("None");
+    const parentTitle = document.querySelector(`.parent-title`);
     parentTitle.value = null;
+    this.filterCategory("None");
     Utils.clearCategories();
   }
 
   filterCategory(category) {
     Utils.showProductsForCategory(category);
-    categoryTitle.value = category;
+    this.categoryTitle.value = category;
     CreepUtils.sortOnCreepiness();
     CreepUtils.moveCreepyFace();
     Utils.checkForEmptyNotice();
@@ -103,7 +161,7 @@ export class SearchFilter {
   }
 
   activateSubcategory(activeClasses, defaultClasses) {
-    const categoryName = categoryTitle.value.trim();
+    const categoryName = this.categoryTitle.value.trim();
 
     document
       .querySelector(`a.subcategories[data-name="${categoryName}"]`)
@@ -123,6 +181,7 @@ export class SearchFilter {
   }
 
   filterSubcategory(category) {
+    const subcategories = document.querySelectorAll(`.subcategories`);
     for (const subcategory of subcategories) {
       if (subcategory.dataset.parent === category) {
         subcategory.classList.remove(`tw-hidden`);
